Do not show header if it's already completely visible
This commit updates `StickyHeaderItemDecorator.kt` to prevent the sticky header from being drawn if the corresponding header item in the list is already fully visible on the screen.

The `findHeaderPosition` method now checks `findFirstCompletelyVisibleItemPosition`. If the identified header position `i` matches this `firstCompletelyVisibleIndex`, it means the header item is fully in view, and thus the sticky header is not needed and `null` is returned.